Contain catfind code

+++ controls/V-72927.rb
@@ -79,13 +79,18 @@ control "V-72927" do
   enabled. To ensure that logging is enabled, review supplementary content
   APPENDIX-C for instructions on enabling logging."
 
-  
+if file(pg_audit_log_dir).exist?  
   describe command("PGPASSWORD='#{pg_dba_password}' psql -U #{pg_dba} -d #{pg_db} -h #{pg_host} -A -t -c \"CREATE ROLE permdeniedtest; SET ROLE permdeniedtest; UPDATE pg_authid SET rolsuper = 't' WHERE rolname = 'permdeniedtest'; DROP ROLE IF EXISTS permdeniedtest;\"") do
     its('stdout') { should match // }
   end
 
   describe command("cat `find #{pg_audit_log_dir} -type f -printf '%T@ %p\n' | sort -n | tail -1 | cut -f2- -d\" \"` | grep \"permission denied for relation pg_authid\"") do
     its('stdout') { should match /^.*permission denied for relation pg_authid.*$/ }
+  end 
+else
+  describe "The #{pg_audit_log_dir} directory was not found. Check path for this postgres version/install to define the value for the 'pg_audit_log_dir' inspec input parameter." do
+    skip "The #{pg_audit_log_dir} directory was not found. Check path for this postgres version/install to define the value for the 'pg_audit_log_dir' inspec input parameter."
   end
+end
 
 end
